refactor: Footer를 Field 공용 컴포넌트로 이전
