+++ packages/jds/src/components/Field/index.tsx
@@ -4,3 +4,4 @@ export type { FieldContentProps } from "./compound/Content";
 export type { FieldLabelProps } from "./compound/Label";
 export type { FieldHelperProps } from "./compound/Helper";
 export type { FieldCounterProps } from "./compound/Counter";
+export type { FieldFooterProps } from "./compound/Footer";
