+++ packages/jds/src/components/MultiSelectField/index.ts
@@ -1,8 +1,8 @@
 export { MultiSelectField } from "./MultiSelectField";
 export type { MultiSelectFieldProps, MultiSelectFieldInputProps } from "./multiSelectField.types";
-export type { MultiSelectFieldFooterProps } from "./compound/Footer";
 export type {
   FieldContentProps as MultiSelectFieldContentProps,
   FieldLabelProps as MultiSelectFieldLabelProps,
   FieldHelperProps as MultiSelectFieldHelperProps,
+  FieldFooterProps as MultiSelectFieldFooterProps,
 } from "../Field";
